requirements.txt
urllib3
requests
lxml
bleach
termcolor
python-irodsclient

